fix usage on qemu boards
- mostly just check for fw_env.config existence before manipulating it
- fstype has to be ext4 for extlinux; might be more flexible in the
future if we have qemu board use u-boot as well
- /boot/uboot_env.d/*'s empty check isn't enough when a README exists,
check for any non-empty non-comment line instead

+++ scripts/post_rootfs.sh
@@ -108,7 +108,7 @@ post_rootfs() {
 ${rootdev}boot${ab} 0x3fe000 0x2000
 ${rootdev}boot${ab} 0x3fa000 0x2000
 EOF
-		else
+		elif [ -e /etc/fw_env.config ]; then
 			cp /etc/fw_env.config /target/etc/fw_env.config \
 				|| error "Could not copy fw_env.config"
 		fi
